Made changes to layout
Added fonts,
componentalized pieces of index page
added ball simulation
added points field
added blur field

general cleanup
