+++ package.json
@@ -3,15 +3,20 @@
   "version": "2.0.1",
   "private": true,
   "dependencies": {
+    "@react-three/cannon": "^6.5.0",
     "@react-three/drei": "^9.43.3",
     "@react-three/fiber": "^8.9.1",
+    "@react-three/postprocessing": "^2.7.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "lerp": "^1.0.3",
+    "nice-color-palettes": "^3.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.3",
     "react-scripts": "5.0.1",
+    "react-use-scroll-snap": "^0.0.4",
     "three": "^0.146.0",
     "three-globe": "^2.24.10",
     "use-asset": "^1.0.4",
